i can deserialize
i think this is jank. i think i can rely completely on serde json if i
figure out a better way to reference functions

i need something for a FunctionCall node to refer to. maybe it's just a
FunctionCallReference node which contains a reference to a unique
function call ID.